requirements.txt
pydantic>=2.6
python-dotenv>=1.0
groq>=0.9.0
rich>=13.7